service/version-api/backend/requirements.txt
cherrypy==18.1.1
pyyaml~=5.4.1
psutil~=5.8.0
torch~=1.8.1
librosa~=0.8.0
numpy~=1.19.4
requests~=2.25.1